Moved Signal to seperate file in base_classes folder, renamed to SignalBase.
